[plumbing] Add basic Hakyll generator
This patch adds the code for a static site generator built on
[Hakyll][hakyll].  Hakyll is flexible enough to do the things that I
need for my website, but doesn't need an interpreter and many runtime
dependencies just to generate the website.

Right now, all the static site generating code does is copy the
existing files that are part of the website into an output directory.

[hakyll]: https://jaspervdj.be/hakyll/

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>