[plumbing] Move index.html to Hakyll generation
Hakyll comes with a templating feature that allows us to write the
markup that is common to each page in one place, and be inserted into
each page during generation.  This patch tests this functionality with
just the `index.html` file, aiming to use the templating functionality
while making no changes to the resulting generated file.

Concerns
------------------------------------------------------------------------

The overall idea is to create a `templates/` directory that stores the
common template files, to add a template containing the scaffold of an
HTML page, and to change the site generation program to apply this
template when building `index.html`.  This allows us to separate the
distinguishing content of the home page from the structure, which will
presumably be shared by other pages later.

There are two main concerns here:

  1. In order to make no changes to the resulting generated file, I
     need the HTML output to be indented as before.  By default,
     though, Hakyll templates just interpolate the file directly into
     the template, making the first line indented, but subsequent
     lines not.  We could require that every non-template file be
     indented as required by the templates, but in addition to not
     being elegant, that solution couples together the content and the
     template in a non-obvious way.  Instead, we add functionality to
     the site generation that adds 𝑛 spaces of indentation to the page
     body before interpolating, resulting in the correct indentation.

  2. [This site’s URIs are cool][cool-uris], and they don’t have file
     extensions.  Instead, when fielding a request to `a/b/foo`, it
     will look for a file `a/b/foo.html` and serve that.  This means,
     in the case of this patch, I need my Hakyll generator to create a
     file `index.html` (viz., to route `index.html` ↦ `index.html`),
     but to convince the file that it is at the URI `/`.  This is done
     with a custom `url` field in the file’s “context”.

[cool-uris]: https://www.w3.org/Provider/Style/URI

Testing
------------------------------------------------------------------------

Because the goal of this patch is to make no difference in the
generated website, any regressions can easily be caught by diffing the
results of building the site with these modifications against the
results of building the site at the previous commit.  This patch
introduces no changes into the generated output.

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>