[backend] Correct `Caddyfile` syntax
The `tls` directive is not allowed at the top level of a `Caddyfile`.
Instead, it has to be scoped within a server block.  This patch moves
the `tls` directive into the server block.

Testing
------------------------------------------------------------------------

This patch incorporates a change that was being used on the deployed
version of pniedzielski.net.

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>