[plumbing] Add a Brotli compression compiler
Our webserver Caddy automatically serves [Brotli][brotli]-compressed
files with the `.br` extension if they exist.  Because Brotli is
supported by major web browsers and Cloudflare, the CDN I use, and
because it is both performant and gives a high compression ratio, it
is a good candidate for automatic generation by this static site
generator.

This patch adds a function that returns a `Compiler` representing how
to compress an `Item` with the Brotli compression algorithm.  As an
implementation-choice, we use the Unix `brotli` command, because it
has a stable command line interface.  There is a [Haskell library that
write the C libraries][hs-brotli], but it warns that it is alpha
quality, and it's not in Hackage yet, so we choose instead to use an
executable as a filter.

The function was manually verified by setting up a temporary version
of `index.html` that used this compiler to compress the existing
`index.html`, and then decompressing the result and `diff`ing against
the original.

[brotli]:    https://www.brotli.org/
[brotli-hs]: https://github.com/iand675/hs-brotli

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>