[layout] Make nav links in default template absolute
This patch changes the relative links in the navigation section of the
default template (which is currently only used by `index.html`) to
absolute links to match the other links on the site.  This will allow
us to reuse the template with other generated pages.

Testing
------------------------------------------------------------------------

To test this patch, build the site at this commit, and verify that the
generated `index.html` uses absolute links in the `<nav>` section
instead of relative links.

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>