[plumbing] Move articles to Hakyll generation
This patch switches the article posts to being generated with Hakyll
templating, similarly to the `index.html` file.  Just as with that
change, this patch aims to use the templating functionality while
making no changes to the resulting generated file.

Testing
------------------------------------------------------------------------

To test this patch, diff the result of generating the site with this
patch against the result of generating the site without this patch and
verify that there are no differences.

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>