[plumbing] Teach the site generator to use Brotli
Now that our static site generator has a working routine to compress
an item using the [Brotli][brotli] compression algorithm, we can teach
it to actually use the routine on all relevant files.  This patch adds
a behavior `brotliBehavior` that generates a Brotli compressed file
next to the original file, with an extra `.br` extension, and adds
this behavior to the following three places:

  - Static files :: Any file that only copied into the new location
    can be Brotli compressed.  This includes CSS files, Javascript
    files, images, and few other resources.  There likely isn't much
    benefit to using Brotli compression on already compressed images,
    but doing this doesn't hurt.  However, because the CSS and
    Javascript are needed for every page, we potentially have a great
    benefit to compressing these.

  - `index.html` :: Because this is just an HTML document, it can
    benefit greatly from using Brotli compression.  This is very
    beneficial, since most hits to an HTML page are to `index.html`.

  - Posts :: These are also HTML documents, and so they can benefit
    from Brotli compression.  However, since there may potentially be
    very many of these, we need to be careful in making sure that the
    cost of compressing them isn't too high.  For now, there is no
    problem.

This patch was verified by spot-checking the resulting files,
decompressing them, and comparing them to the original file.

[brotli]:    https://www.brotli.org/

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>