[plumbing] Import an initial Jekyll site
[Jekyll][jekyll] is a static site generator written in Ruby.  I have
used Jekyll in the past, before I toyed around with [Hakyll][hakyll].
Although the latter was nice, I found it got in my way too much,
forcing me to roll too much code of my own to get what I wanted, and
this ultimately increased the barrier for posting things on my
website.  Because of this, I have decided to switch back to Jekyll.

This commit adds the default generated Jekyll site, created as
described on the [Using Jekyll with Bundler][jekyll-bundler] page on
the Jekyll website.  Specifically, the following commands got us this
site:

    bundle init
    bundle install --path _vendor/bundle
    bundle add jekyll
    bundle update listen
    bundle add jekyll
    bundle exec jekyll new --force --skip-bundle .

Additionally, I cleaned up the `.gitignore` to ignore the
`_vendor/bundle` directory and the `Gemfile.lock` file.

[jekyll]:         https://jekyllrb.com/
[hakyll]:         https://jaspervdj.be/hakyll/
[jekyll-bundler]: https://jekyllrb.com/tutorials/using-jekyll-with-bundler/

Testing
=======

To test this commit, run

    bundle exec jekyll serve

and navigate to `http://localhost:4000/` in a web browser.  Make sure
that the website functions as expected.

License: CC-BY4.0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>