[plumbing] Import `minima` theme
By default, Jekyll has started using Ruby Gems for themes.  The Jekyll
default theme is called `minima`.  I don’t need this level of
flexibility, and in fact in my use case, it’s best that the layout be
in the same repository as the rest of the website.  This patch copies
the layout from `minima` to my website’s repository.

License: CC-BY4.0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>