[plumbing] Automatically Brotli/Gzip compress files
Our webserver Caddy automatically serves [Brotli][brotli]-compressed
files with the `.br` extension and [Gzip][gzip]-compressed files with
the `.gz` extension if either exists.  Because Brotli is supported by
major web browsers and Cloudflare, the CDN I use, and because it is
both performant and gives a high compression ratio, it is a good
candidate for automatic generation by this static site generator.  As
a fall back, I provide Gzip files, which browsers have supported
forever.

This patch enables automatic compression of all files generated by the
static site generator with both Brotli and Gzip.

There are two Ruby gems that enable this:

  - `jekyll-brotli`
  - `jekyll-gzip`

This patch was manually verified by generating the site, and making
sure that each file had a corresponding file with an additional `.br`
extension for Brotli compression, and that each file had a
corresponding file with an additional `.gzip` extension for Gzip
compression.

[brotli]:    https://www.brotli.org/
[gzip]:      https://www.gzip.org/

License: CC-BY4.0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>