[plumbing] Ignore `stack.yaml.lock`.
Lock files do not belong in a repository.  Used this way, they are a
poor man’s version of reproducible builds, and don’t actually serve to
ensure the environment is fully reproducible.  This patch makes sure
we don’t accidentally commit it by adding it to our `.gitignore` file.

License: AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>