[content] Change my picture
The previous patch (`1e9837d`) makes my picture visible on my web
page, correcting a long-standing bug.  The picture on my web site is
quite old, and low resolution, though.  This patch updates the picture
to a more new and higher resolution image, taken when I was at
Cornell.

Testing
-------------------------------------------------------------------------------

This patch was manually verified, by generating the site, loading it
in a web browser (bypassing caching), and ensuring that the new image
loads.

License:       AGPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>